fix(backend): widen multiplication_factor from INT to BIGINT to prevent overflow
For salted hash types, multiplication_factor = rule_count × salt_count which
easily exceeds INT max (2,147,483,647). Example: 48,414 rules × 82,780 salts
= 4,008,069,004 caused "pq: value out of range for type integer" on job creation.

- Add migration 000129 to ALTER both job_executions and preset_jobs columns
- Change Go model fields from int to int64
- Remove int() truncation casts across services, repositories, and handlers
- Fix type comparisons with *maxRuleEnd at scheduling/cleanup/progress sites

+++ backend/internal/handlers/jobs/user_jobs.go
@@ -115,7 +115,7 @@ type JobSummary struct {
 	CreatedByUsername      *string `json:"created_by_username,omitempty"`
 	ErrorMessage           *string `json:"error_message,omitempty"`
 	EffectiveKeyspace      *int64  `json:"effective_keyspace,omitempty"`
-	MultiplicationFactor   int     `json:"multiplication_factor,omitempty"`
+	MultiplicationFactor   int64   `json:"multiplication_factor,omitempty"`
 	UsesRuleSplitting      bool    `json:"uses_rule_splitting"`
 	ProcessedKeyspace      *int64  `json:"processed_keyspace,omitempty"`
 	DispatchedKeyspace     *int64  `json:"dispatched_keyspace,omitempty"`
